+++ examples/MincWorld/Tmaze.cpp
@@ -18,7 +18,7 @@ bool Tmaze::IsDuplicate(Tmaze* maze) {
 	if ((int)path.GetCount() != (int)maze->path.GetCount())
 		return false;
 
-	for (int i = 0; i < (int)path.GetCount(); i++) {
+	for (int i = 0; i < path.GetCount(); i++) {
 		if (path[i].mark != maze->path[i].mark)
 			return false;
